requirements.txt
streamlit>=1.33
python-dotenv>=1.0.1
python-docx>=1.1.2
google-genai>=0.3.0
